+++ AM.Application.Shell/ServiceCollectionExtensions.cs
@@ -36,6 +36,7 @@ using AM.Modules.Alarm;
 using AM.Modules.Dashboard;
 using AM.Modules.Identity;
 using AM.Modules.IoMonitor;
+using AM.Modules.Engineering;
 using AM.Modules.Motion;
 using AM.Modules.Parameter;
 using AM.Services;
@@ -134,6 +135,7 @@ internal static class ServiceCollectionExtensions
         services.AddSingleton<IdentityViewModel>();
         services.AddSingleton<MotionViewModel>();
         services.AddSingleton<ParameterViewModel>();
+        services.AddSingleton<EngineeringViewModel>();
         services.AddSingleton<ShellViewModel>(); // header + alarm bar + connection chips
         return services;
     }